Move recognition module under lib/ and update imports
